trader/requirements.txt
requests>=2.31.0
websocket-client>=1.6.1
pandas>=2.0.3
numpy>=1.24.3
python-dotenv>=1.0.0
schedule>=1.2.0
pydantic>=2.0.0
aiohttp>=3.8.5
asyncio-mqtt>=0.16.1
ta-lib>=0.4.28
plotly>=5.15.0
dash>=2.10.1
sqlalchemy>=2.0.19
python-telegram-bot>=20.4
loguru>=0.7.0
pytz>=2023.3
